+++ Modules/SharePointDsc/DSCResources/MSFT_SPFarm/MSFT_SPFarm.psm1
@@ -80,18 +80,25 @@ function Get-TargetResource
             Write-Verbose -Message "Detected installation of SharePoint 2013"
         }
         16 {
-            Write-Verbose -Message "Detected installation of SharePoint 2016"
+            if($installedVersion.ProductBuildPart.ToString().Length -eq 4)
+            {
+                Write-Verbose -Message "Detected installation of SharePoint 2016"
+            }
+            else
+            {
+                Write-Verbose -Message "Detected installation of SharePoint 2019"
+            }
         }
         default {
             throw ("Detected an unsupported major version of SharePoint. SharePointDsc only " + `
-                   "supports SharePoint 2013 or 2016.")
+                   "supports SharePoint 2013, 2016 or 2019.")
         }
     }
 
     if (($PSBoundParameters.ContainsKey("ServerRole") -eq $true) `
         -and $installedVersion.FileMajorPart -ne 16)
     {
-        throw [Exception] "Server role is only supported in SharePoint 2016."
+        throw [Exception] "Server role is only supported in SharePoint 2016 and 2019."
     }
 
     if (($PSBoundParameters.ContainsKey("ServerRole") -eq $true) `
@@ -418,6 +425,13 @@ function Set-TargetResource
 
         $modulePath = "..\..\Modules\SharePointDsc.Farm\SPFarm.psm1"
         Import-Module -Name (Join-Path -Path $scriptRoot -ChildPath $modulePath -Resolve)
+        $sqlInstanceStatus = Get-SPDSCSQLInstanceStatus -SQLServer $params.DatabaseServer `
+
+        if ($sqlInstanceStatus.MaxDOPCorrect -ne $true)
+        {
+            throw "The MaxDOP setting is incorrect. Please correct before continuing."
+        }
+
         $dbStatus = Get-SPDSCConfigDBStatus -SQLServer $params.DatabaseServer `
                                             -Database $params.FarmConfigDatabaseName
 
@@ -444,7 +458,8 @@ function Set-TargetResource
             SkipRegisterAsDistributedCacheHost = $true
         }
 
-        switch((Get-SPDSCInstalledProductVersion).FileMajorPart)
+        $installedVersion = Get-SPDSCInstalledProductVersion
+        switch($installedVersion.FileMajorPart)
         {
             15 {
                 Write-Verbose -Message "Detected Version: SharePoint 2013"
@@ -452,22 +467,39 @@ function Set-TargetResource
             16 {
                 if ($params.ContainsKey("ServerRole") -eq $true)
                 {
-                    Write-Verbose -Message ("Detected Version: SharePoint 2016 - " + `
-                                            "configuring server as $($params.ServerRole)")
+                    if($installedVersion.ProductBuildPart.ToString().Length -eq 4)
+                    {
+                        Write-Verbose -Message ("Detected Version: SharePoint 2016 - " + `
+                                                "configuring server as $($params.ServerRole)")
+                    }
+                    else
+                    {
+                        Write-Verbose -Message ("Detected Version: SharePoint 2019 - " + `
+                                                "configuring server as $($params.ServerRole)")
+                    }
                     $executeArgs.Add("LocalServerRole", $params.ServerRole)
                 }
                 else
                 {
-                    Write-Verbose -Message ("Detected Version: SharePoint 2016 - no server " + `
-                                            "role provided, configuring server without a " + `
-                                            "specific role")
+                    if($installedVersion.ProductBuildPart.ToString().Length -eq 4)
+                    {
+                        Write-Verbose -Message ("Detected Version: SharePoint 2016 - no server " + `
+                                                "role provided, configuring server without a " + `
+                                                "specific role")
+                    }
+                    else
+                    {
+                        Write-Verbose -Message ("Detected Version: SharePoint 2019 - no server " + `
+                                                "role provided, configuring server without a " + `
+                                                "specific role")
+                    }
                     $executeArgs.Add("ServerRoleOptional", $true)
                 }
             }
             Default {
                 throw [Exception] ("An unknown version of SharePoint (Major version $_) " + `
-                                    "was detected. Only versions 15 (SharePoint 2013) or " + `
-                                    "16 (SharePoint 2016) are supported.")
+                                    "was detected. Only versions 15 (SharePoint 2013) and" + `
+                                    "16 (SharePoint 2016 or SharePoint 2019) are supported.")
             }
         }