+++ Tests/Unit/SharePointDsc/SharePointDsc.SPFarm.Tests.ps1
@@ -43,9 +43,17 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
         Mock -CommandName "Start-Service" -MockWith { }
         Mock -CommandName "Stop-Service" -MockWith { }
         Mock -CommandName "Start-SPServiceInstance" -MockWith { }
+        Mock -CommandName Get-SPDSCInstalledProductVersion {
+            return @{
+                FileMajorPart = $Global:SPDscHelper.CurrentStubBuildNumber.Major
+                FileBuildPart = $Global:SPDscHelper.CurrentStubBuildNumber.Build
+                ProductBuildPart = $Global:SPDscHelper.CurrentStubBuildNumber.Build
+            }
+        }
+
 
         # Test Contexts
-        Context -Name "No config databaes exists, and this server should be connected to one" -Fixture {
+        Context -Name "No config databases exists, and this server should be connected to one" -Fixture {
             $testParams = @{
                 IsSingleInstance = "Yes"
                 Ensure = "Present"
@@ -64,6 +72,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     Locked = $false
                     ValidPermissions = $true
                     DatabaseExists = $false
+                }
+            }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
                 }
             }
             Mock -CommandName "Get-SPWebApplication" -MockWith {
@@ -88,7 +101,7 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
             }
         }
 
-        Context -Name "No config databaes exists, and this server should be connected to one but won't run central admin" -Fixture {
+        Context -Name "No config databases exists, and this server should be connected to one but won't run central admin" -Fixture {
             $testParams = @{
                 IsSingleInstance = "Yes"
                 Ensure = "Present"
@@ -107,6 +120,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     Locked = $false
                     ValidPermissions = $true
                     DatabaseExists = $false
+                }
+            }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
                 }
             }
             Mock -CommandName "Get-SPWebApplication" -MockWith {
@@ -149,6 +167,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     Locked = $false
                     ValidPermissions = $true
                     DatabaseExists = $true
+                }
+            }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
                 }
             }
             Mock -CommandName "Get-SPWebApplication" -MockWith {
@@ -220,6 +243,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     Locked = $false
                     ValidPermissions = $true
                     DatabaseExists = $true
+                }
+            }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
                 }
             }
             Mock -CommandName "Get-SPWebApplication" -MockWith {
@@ -301,6 +329,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     }
                 }
             }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
+                }
+            }
             Mock -CommandName "Get-SPWebApplication" -MockWith {
                 return @{
                     IsAdministrationWebApplication = $true
@@ -326,6 +359,7 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
 
         Context -Name "Server is connected to farm, but Central Admin isn't started" -Fixture {
             $testParams = @{
+                IsSingleInstance = "Yes"
                 Ensure = "Present"
                 FarmConfigDatabaseName = "SP_Config"
                 DatabaseServer = "sql.contoso.com"
@@ -353,6 +387,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     Locked = $false
                     ValidPermissions = $true
                     DatabaseExists = $true
+                }
+            }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
                 }
             }
             Mock -CommandName Get-SPDatabase -MockWith {
@@ -418,6 +457,7 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
 
         Context -Name "This server is connected to the farm and is running CA, but shouldn't" -Fixture {
             $testParams = @{
+                IsSingleInstance = "Yes"
                 Ensure = "Present"
                 FarmConfigDatabaseName = "SP_Config"
                 DatabaseServer = "sql.contoso.com"
@@ -445,6 +485,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     Locked = $false
                     ValidPermissions = $true
                     DatabaseExists = $true
+                }
+            }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
                 }
             }
             Mock -CommandName "Get-SPDatabase" -MockWith {
@@ -534,6 +579,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     Locked = $false
                     ValidPermissions = $true
                     DatabaseExists = $true
+                }
+            }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
                 }
             }
             Mock -CommandName "Get-SPDatabase" -MockWith {
@@ -666,6 +716,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                         DatabaseExists = $true
                     }
                 }
+                Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                    return @{
+                        MaxDOPCorrect = $true
+                    }
+                }
                 Mock -CommandName "Get-SPDatabase" -MockWith {
                     return @(@{
                         Name = $testParams.FarmConfigDatabaseName
@@ -762,17 +817,15 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                         DatabaseExists = $false
                     }
                 }
+                Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                    return @{
+                        MaxDOPCorrect = $true
+                    }
+                }
                 Mock -CommandName "Get-SPWebApplication" -MockWith {
                     return @{
                         IsAdministrationWebApplication = $true
                         Url = "http://localhost:12345"
-                    }
-                }
-
-                Mock -CommandName Get-SPDSCInstalledProductVersion -MockWith {
-                    return @{
-                        FileMajorPart = 16
-                        FileBuildPart = 4456
                     }
                 }
 
@@ -821,6 +874,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     DatabaseExists = $true
                 }
             }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
+                }
+            }
             Mock -CommandName "Get-SPDatabase" -MockWith {
                 return @(@{
                     Name = $testParams.FarmConfigDatabaseName
@@ -849,7 +907,7 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                 }
             }
 
-            Mock -CommandName Get-SPDSCInstalledProductVersion -MockWith { return @{ FileMajorPart = 16 } }
+            Mock -CommandName Get-SPDSCInstalledProductVersion -MockWith { return @{ FileMajorPart = 16; ProductBuildPart = 4700 } }
 
             It "Should return WebFrontEnd from the get method"{
                 (Get-TargetResource @testParams).ServerRole | Should Be "WebFrontEnd"
@@ -899,6 +957,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     Locked = $false
                     ValidPermissions = $false
                     DatabaseExists = $false
+                }
+            }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
                 }
             }
             Mock -CommandName "Get-SPDatabase" -MockWith {
@@ -974,6 +1037,11 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
                     Locked = $false
                     ValidPermissions = $true
                     DatabaseExists = $true
+                }
+            }
+            Mock -CommandName "Get-SPDSCSQLInstanceStatus" -MockWith {
+                return @{
+                    MaxDOPCorrect = $true
                 }
             }
             Mock -CommandName "Get-SPDatabase" -MockWith {