+++ Tests/Unit/UnitTestHelper.psm1
@@ -45,14 +45,13 @@ function New-SPDscUnitTestHelper
     }
 
     $spBuild = (Get-Item -Path $SharePointStubModule).Directory.BaseName
-    $firstDot = $spBuild.IndexOf(".")
-    $majorBuildNumber = $spBuild.Substring(0, $firstDot)
+    $spBuildParts = $spBuild.Split('.')
+    $majorBuildNumber = $spBuildParts[0]
+    $minorBuildNumber = $spBuildParts[1]
 
     $describeHeader += " [SP Build: $spBuild]"
 
     Import-Module -Name $moduleToLoad -Global
-
-
 
     $initScript = @"
             Remove-Module -Name "Microsoft.SharePoint.PowerShell" -Force -ErrorAction SilentlyContinue
@@ -67,6 +66,10 @@ function New-SPDscUnitTestHelper
 
             Mock -CommandName Get-SPDSCAssemblyVersion -MockWith {
                 return $majorBuildNumber
+            }
+
+            Mock -CommandName Get-SPDSCBuildVersion -MockWith {
+                return $minorBuildNumber
             }
 
 "@