+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change log for SharePointDsc
 
+## Unreleased
+
+* SPDatabaseAAG
+  * Updated readme.md to specify that this resource also updates the database
+    connection string
+* SPDiagnosticsProvider
+  * Fixed issue where enabling providers did not work
+* SPLogLevel
+  * Added High as TraceLevel, which was not included yet
+* SPSearchServiceApp
+  * Updated Set method to check if service application pool exists. Resource
+    will throw an error if it does not exist
+* SPRemoteFarmTrust
+  * Updated readme.md file to add a link that was lost during earlier updates
+* SPSearchTopology
+  * Fixed issue where Get method threw an error when the specified service
+    application didn't exist yet
+  * Fixed issue where the resource would fail is the FQDN was specified
+* SPTrustedIdentityTokenIssuerProviderRealm
+  * Fixed issue where Get method threw an error when the realm didn't exist yet
+
 ## 2.4
 
 * SPCacheAccounts